+++ src/components/Cards.tsx
@@ -3,7 +3,7 @@ import { SemanticWIDTHS, Grid } from 'semantic-ui-react';
 import styled from 'styled-components';
 import { Card as CardType } from '../store/models';
 import Card from './Card';
-import { HeadingLevel } from './Heading';
+import Heading, { HeadingLevel } from './Heading';
 
 const Container = styled.div`
   margin-top: ${p => p.theme.spacing.xl};
@@ -11,14 +11,21 @@ const Container = styled.div`
 `;
 
 interface Props {
+  title?: string | null;
   cards: CardType[];
   columns?: SemanticWIDTHS;
   headingLevel?: HeadingLevel;
 }
 
-const Cards: React.FC<Props> = ({ cards, columns = 4, headingLevel }) => {
+const Cards: React.FC<Props> = ({
+  title,
+  cards,
+  columns = 4,
+  headingLevel,
+}) => {
   return (
     <Container>
+      {title && <Heading level="h2">{title}</Heading>}
       <Grid columns={columns} stackable doubling centered stretched>
         {cards?.map(card => (
           <Grid.Column key={card.id}>