+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - OPI-107: View for user's profile (including change password & delete account)
+- OPI-106: Cards can have title field
 
 ### Fixed